requirements.txt
cycler==0.10.0
dataclasses==0.8
dnspython==2.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.2.0
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.16.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
websockets==9.0.1
